Update MASM and Facial recognition to Python 3.8 and be more reliable

+++ Additions_FaceDetectionRecognition/MASM/scripts/numpy/lib/_datasource.py
@@ -41,7 +41,11 @@ import warnings
 import shutil
 import io
 
+from numpy.core.overrides import set_module
+
+
 _open = open
+
 
 def _check_mode(mode, encoding, newline):
     """Check mode and that encoding and newline are compatible.
@@ -262,7 +266,8 @@ def open(path, mode='r', destpath=os.curdir, encoding=None, newline=None):
     return ds.open(path, mode, encoding=encoding, newline=newline)
 
 
-class DataSource (object):
+@set_module('numpy')
+class DataSource(object):
     """
     DataSource(destpath='.')
 
@@ -323,7 +328,7 @@ class DataSource (object):
 
     def __del__(self):
         # Remove temp directories
-        if self._istmpdest:
+        if hasattr(self, '_istmpdest') and self._istmpdest:
             shutil.rmtree(self._destpath)
 
     def _iszip(self, filename):
@@ -540,6 +545,11 @@ class DataSource (object):
         is accessible if it exists in either location.
 
         """
+
+        # First test for local path
+        if os.path.exists(path):
+            return True
+
         # We import this here because importing urllib2 is slow and
         # a significant fraction of numpy's total import time.
         if sys.version_info[0] >= 3:
@@ -548,10 +558,6 @@ class DataSource (object):
         else:
             from urllib2 import urlopen
             from urllib2 import URLError
-
-        # Test local path
-        if os.path.exists(path):
-            return True
 
         # Test cached url
         upath = self.abspath(path)
